feat: add core module
