+++ package.json
@@ -4,18 +4,27 @@
   "description": "Obfuscating text or phrases with random uncommon characters to avoid banning.",
   "type": "module",
   "main": "dist/index.js",
-  "types": "dist/index.d.ts",
   "files": [
     "charsets/",
     "dist/"
   ],
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./core": {
+      "types": "./dist/core.d.ts",
+      "default": "./dist/core.js"
+    }
+  },
   "scripts": {
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:cov": "vitest run --coverage",
-    "build": "tsup src/index.ts --format esm -d dist --clean --dts src/index.ts --sourcemap --minify",
+    "build": "tsup src/index.ts src/core.ts --format esm -d dist --clean --dts --sourcemap --minify",
     "prepublish": "npm run build"
   },
   "repository": {
